requirements.txt
anyio
asyncpg
click
colorama
databases
fastapi
greenlet
gunicorn
h11
idna
pydantic
python-dotenv
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
psycopg2-binary